fixed bug with 32bit integers on 64 bit machines. Reduced version number strings to one single place. Now is V 0.5.2

+++ pyRserve/rtypes.py
@@ -240,7 +240,7 @@ DT_Map = {
 numpyMap = {
     XT_ARRAY_BOOL:     numpy.bool_,
 #    XT_BYTE:           numpy.byte,
-    XT_ARRAY_INT:      numpy.int,
+    XT_ARRAY_INT:      numpy.int32,
     XT_ARRAY_DOUBLE:   numpy.double,     # double float64
     XT_ARRAY_CPLX:     numpy.complex,
     XT_ARRAY_STR:      numpy.string_,
